update 2 files, delete 1 file and create 3 files
