+++ README.md
@@ -32,6 +32,7 @@ flowchart LR
 - [x] Add Architecture diagram - 16/Oct/2024
 - [x] Skeleton CRUD API - 17/Oct/2024
 - [x] Spring JPA Data - 18/Oct/2024
+- [x] CRUD API | Part 1 - 19/Oct/2024
 - [ ] CRUD API
 - [ ] Improve Architecture Diagram
 - [ ] Spring Data Rest for DB
