+++ src/main/java/io/d4rkr0n1n/sample/Notes.java
@@ -1,0 +1,21 @@
+package io.d4rkr0n1n.sample;
+
+import java.util.UUID;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Table
+@Entity
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Notes {
+  @Id UUID id;
+  String name;
+
+}
