+++ src/main/java/io/d4rkr0n1n/sample/SampleRepository.java
@@ -1,0 +1,11 @@
+package io.d4rkr0n1n.sample;
+
+import java.util.UUID;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SampleRepository extends CrudRepository<Notes, UUID> {
+
+}
